// src/lib/types.ts
import { RequestHandler } from "./request/request-handler";

export enum HTTP_CODES {
    OK = 200,
    CREATED = 201,
    BAD_REQUEST = 400,
    UNAUTHORIZED = 401,
    NOT_fOUND = 404,
    INTERNAL_SERVER_ERROR = 500
}

export enum HTTP_METHODS {
    GET = 'GET',
    POST = 'POST',
    PUT = 'PUT',
    DELETE = 'DELETE',
    OPTIONS = 'OPTIONS'
}


export interface IRoute {
    method: HTTP_METHODS,
    path: string,
    requestHandler: RequestHandler
}